requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pytz==2023.3
python-dateutil==2.8.2
